+++ src/components/table/mobile/table_header_mobile.tsx
@@ -1,0 +1,15 @@
+import React, { FunctionComponent, HTMLAttributes } from 'react';
+import classNames from 'classnames';
+import { CommonProps } from '../../common';
+
+export const EuiTableHeaderMobile: FunctionComponent<
+  CommonProps & HTMLAttributes<HTMLDivElement>
+> = ({ children, className, ...rest }) => {
+  const classes = classNames('euiTableHeaderMobile', className);
+
+  return (
+    <div className={classes} {...rest}>
+      {children}
+    </div>
+  );
+};